Add edit profile to Profile screen, fix update service

+++ backend/src/main/java/com/cafe/backend/dto/ReviewDTO.java
@@ -10,7 +10,7 @@ public record ReviewDTO(
     Long id,
     String title,
     String body,
-    Byte rating,
+    Integer rating,
     LocalDateTime createdAt,
     Long userId,
     Long cafeteriaId
